big changes; moved most of inaturalist_api into new controllers; added place autocomplete
