+++ lib/inaturalist_map_server.js
@@ -2,7 +2,8 @@ var ElasticMapper = require( "elasticmaps" ),
     util = require( "./util" ),
     MapStyles = require( "./inaturalist_map_styles" ),
     esClient = require( "./es_client" ),
-    InaturalistAPI = require( "./inaturalist_api" );
+    InaturalistAPI = require( "./inaturalist_api" ),
+    ObservationsController = require( "./controllers/v1/observations_controller" );
 
 var InaturalistMapserver = { };
 
@@ -49,7 +50,7 @@ InaturalistMapserver.prepareQuery = function( req, callback ) {
     } else if( req.params.style === "places" ) {
       req.postgis.query = InaturalistMapserver.placesQuery( req );
     } else {
-      return callback( { message: "unknown style: " +
+      return callback( { error: "unknown style: " +
         req.params.style, status: 404 }, req );
     }
     return callback( null, req );
@@ -58,7 +59,7 @@ InaturalistMapserver.prepareQuery = function( req, callback ) {
   req.elastic_query.fields = InaturalistAPI.defaultMapFields( );
   // requiring mappable=true to be included in map tiles
   req.query.mappable = "true";
-  InaturalistAPI.reqToElasticQuery( req, function( err, query ) {
+  ObservationsController.reqToElasticQuery( req, function( err, query ) {
     var elastic_query = query;
     var searchHash = esClient.searchHash( elastic_query );
     req.elastic_query.query = searchHash.query;
@@ -72,7 +73,7 @@ InaturalistMapserver.prepareQuery = function( req, callback ) {
         req.elastic_query.aggregations = ElasticMapper.geohashAggregation( req );
         break;
       default:
-        return callback( { message: "unknown style: " +
+        return callback( { error: "unknown style: " +
           req.params.style, status: 404 }, req );
     }
     callback( null, req );
