+++ lib/models/place.js
@@ -13,7 +13,7 @@ Place = function( attrs ) {
 
 Place.findByID = function( id, callback ) {
   if( !parseInt( id ) ) {
-    return callback({ messsage: "invalid place_id", status: "422" });
+    return callback({ messsage: "invalid place_id", status: 422 });
   }
   var cacheKey = "places." + id;
   var place = memoryCache.get( cacheKey );
