+++ lib/models/taxon.js
@@ -116,7 +116,7 @@ Taxon.prototype.prepareForResponse = function( localeOpts ) {
 
 Taxon.findByID = function( id, callback ) {
   if( !parseInt( id ) ) {
-    return callback({ messsage: "invalid taxon_id", status: "422" });
+    return callback({ messsage: "invalid taxon_id", status: 422 });
   }
   var cacheKey = "taxa." + id;
   var taxon = memoryCache.get( cacheKey );
