+++ lib/routes_v1.js
@@ -1,7 +1,9 @@
-var _ = require( "underscore" ),
-    YAML = require( "yamljs" ),
+var YAML = require( "yamljs" ),
     ejs = require( "ejs" ),
     InaturalistAPI = require( "./inaturalist_api" ),
+    ObservationsController = require( "./controllers/v1/observations_controller" ),
+    PlacesController = require( "./controllers/v1/places_controller" ),
+    TaxaContoller = require( "./controllers/v1/taxa_controller" ),
     util = require( "./util" ),
     routesV1 = { };
 
@@ -28,28 +30,28 @@ routesV1.swaggerJSON = function( req, res ) {
 
 routesV1.observations_index = function( req, res ) {
   req.query.ttl = req.query.ttl || 300;
-  InaturalistAPI.observationsIndex( req, function( err, data ) {
+  ObservationsController.search( req, function( err, data ) {
     util.renderJSONPresponse( req, res, err, data );
   });
 };
 
 routesV1.observations_identifiers = function( req, res ) {
   req.query.ttl = req.query.ttl || 300;
-  InaturalistAPI.observationsIdentifiers( req, function( err, data ) {
+  ObservationsController.identifiers( req, function( err, data ) {
     util.renderJSONPresponse( req, res, err, data );
   });
 };
 
 routesV1.observations_observers = function( req, res ) {
   req.query.ttl = req.query.ttl || 300;
-  InaturalistAPI.observationsObservers( req, function( err, data ) {
+  ObservationsController.observers( req, function( err, data ) {
     util.renderJSONPresponse( req, res, err, data );
   });
 };
 
 routesV1.species_counts = function( req, res ) {
   req.query.ttl = req.query.ttl || 300;
-  InaturalistAPI.leafTaxaCounts( req, function( err, data ) {
+  ObservationsController.speciesCounts( req, function( err, data ) {
     util.renderJSONPresponse( req, res, err, data );
   });
 };
@@ -57,15 +59,14 @@ routesV1.species_counts = function( req, res ) {
 routesV1.observations_show = function( req, res ) {
   var err = InaturalistAPI.methodValidationError( req );
   if( err ) { return util.renderError( err, res ); }
-  var params = { query: _.extend( { }, { id: req.params.id }, req.query ) };
-  InaturalistAPI.observationsIndex( params, function( err, data ) {
+  ObservationsController.show( req, function( err, data ) {
     util.renderJSONPresponse( req, res, err, data );
   });
 };
 
 routesV1.taxa_autocomplete = function( req, res ) {
   req.query.ttl = req.query.ttl || 120;
-  InaturalistAPI.taxaAutocomplete( req, function( err, data ) {
+  TaxaContoller.autocomplete( req, function( err, data ) {
     util.renderJSONPresponse( req, res, err, data );
   });
 };
@@ -73,16 +74,21 @@ routesV1.taxa_autocomplete = function( req, res ) {
 routesV1.taxa_show = function( req, res ) {
   var err = InaturalistAPI.methodValidationError( req );
   if( err ) { return util.renderError( err, res ); }
-  var ids = _.filter( req.params.id.split(","), _.identity );
-  var filters = [{ terms: { id: ids } }];
-  InaturalistAPI.taxaSearchQuery( req, null, filters, function( err, data ) {
+  TaxaContoller.show( req, function( err, data ) {
+    util.renderJSONPresponse( req, res, err, data );
+  });
+};
+
+routesV1.places_autocomplete = function( req, res ) {
+  req.query.ttl = req.query.ttl || 300;
+  PlacesController.autocomplete( req, function( err, data ) {
     util.renderJSONPresponse( req, res, err, data );
   });
 };
 
 routesV1.places_nearby = function( req, res ) {
   req.query.ttl = req.query.ttl || 300;
-  InaturalistAPI.placesNearby( req, function( err, data ) {
+  PlacesController.nearby( req, function( err, data ) {
     util.renderJSONPresponse( req, res, err, data );
   });
 };
@@ -90,7 +96,7 @@ routesV1.places_nearby = function( req, res ) {
 routesV1.places_show = function( req, res ) {
   var err = InaturalistAPI.methodValidationError( req );
   if( err ) { return util.renderError( err, res ); }
-  InaturalistAPI.placesShow( req, function( err, data ) {
+  PlacesController.show( req, function( err, data ) {
     util.renderJSONPresponse( req, res, err, data );
   });
 };
