+++ lib/util.js
@@ -6,13 +6,10 @@ var _ = require( "underscore" ),
 
 util.renderError = function( err, res ) {
   util.debug( err );
-  var message = ( config.debug && err.message ) ? err.message : "Error";
-  util.renderMessage( res, message, err.status || 500 );
-};
-
-util.renderMessage = function( res, message, status ) {
-  res.set( "Content-Type", "text/html" );
-  res.status( status ).send( message ).end( );
+  var status = err.status ? err.status : 500;
+  var message = ( config.debug && err.error ) ? err.error : "Error";
+  res.status( status );
+  res.jsonp({ error: message, status: status });
 };
 
 // express middleware which logs request info and duration
@@ -26,6 +23,19 @@ util.timingMiddleware = function( req, res, next ) {
     util.debug( logText );
   });
   next( );
+};
+
+util.localeOpts = function( req ) {
+  var localeOpts = { };
+  if( req.query.locale ) { localeOpts.locale = req.query.locale; }
+  localeOpts.locale = localeOpts.locale || "en";
+  if( req.inat && req.inat.preferredPlace ) {
+    localeOpts.preferredPlace = req.inat.preferredPlace;
+  }
+  if( req.inat && req.inat.place ) {
+    localeOpts.place = req.inat.place;
+  }
+  return localeOpts;
 };
 
 // express middleware which sets access control
